Updated openstack/openstack Project: openstack/swift  410b48dfa5e89b325866280225488a53bf564781
SLO docs(tring): manifest params in table, code block; copyedits

It's æsthetically pleasing to display the segment description parameters
in a table (using the reStructuredText "simple" table markup) and to use
a code block (which receives syntax hightlighting in the rendered
document!) for the example JSON fragment. Also, a few very minor copy
changes are made (a comma after an introductory phrase clarifies; JSON
and ETag receieve their canonical capitalizations; the elements of the
manifest JSON aren't "files" exactly; the inclusive nature of the range
is noted for the benefit of harried readers who won't read the "Range
Specification" section later in the document).

Change-Id: I5296fc04c6fe5410ad14bba7de0a7f8ab060b126