Updated openstack/openstack Project: openstack/nova  7593ca606468561d8612a759009ab10dd10e9eb7
libvirt:on snapshot delete, use qemu-img to blockRebase if VM is stopped

Libvirt won't do a blockRebase on a domain that is not running. So,
in that case, use "qemu-img rebase" instead.

Note: For now, trying to rebase a network disk using qemu-img raises
a NovaException error because I can't test that it successfully works
for every protocol (gluster, sheepdog, etc) that executes this code
path. I successfully tested this with file-based disk.

Change-Id: I0e6819a6c8dc70b9bd7d1a9c18dc185b4537a3e4
Closes-Bug: #1444806
Closes-Bug: #1465416