Updated openstack/openstack Project: openstack-infra/storyboard  428bd522f20951e0f2a5e9b6b12fe4561e213e97
Add API functionality for automatic worklist filters

Automatic worklists are worklists whose contents are automatically
generated on the fly based on filters set by the owners of the worklist.

This can be used to create a list which automatically updates its
contents, eg. "All the unmerged tasks assigned to Adam Coldrick".

Change-Id: I5ea21808cc5aad058b1309e22d22f409b09b5c0a