Updated openstack/openstack Project: openstack/ironic  178d9a4a2e112be6218a3445c217017a42467b3b
iPXE document about the existence of prebuilt images

Prior to this patch the Ironic documentation was only mentioning about
the images which are packaged by the supported distros and about compiling
an iPXE image from source.  This patch extends the documentation to talk
about the prebuilt images that are distributed by the iPXE project.

The patch also fixes the syntax of the note RST tag.

Change-Id: I3b4230e5a43e4db5268e1d0c2a1dcdf25ffac396