Updated openstack/openstack Project: openstack/barbican  4ffb4cf569d2fb946a8a31261d6e632705e9232c
Fix Intermittant Unit Test Failure in Repositories

When sql_connection is null, then migration.init_config() will bail
with a run time error.  The default config value is null, so that
value needs to be overridden before migration.init_config() is called.
This usually works, either because an earlier test has overridden the
value or the test finds a barbican.conf file.  But, sometimes, it will
fail.

Change-Id: I5ffbcd037d7fd4d8841571c834378ecdf1d8d454