Updated openstack/openstack Project: openstack-dev/grenade  19de9d8b09da7fa034eb4d7aa01c91087ae229b8
always set upgrade_level to auto

The partial-ncpu job needed to do a very special dance to set, only on
the partial-ncpu jobs, the upgrade_level, before the new side comes
up.

Now that nova supports the 'auto' stanza, we should just set it all
the time on upgrade. Nodes that get upgraded will know to be auto,
nodes that are not will end up using their normal default latest.

Depends-On: Ieb92f52180369f2513bda40dcbaa9db392816841

Change-Id: I5c899779866ebdf492d474a625eb396273bb0259