Updated openstack/openstack Project: openstack/python-novaclient  bba1d6ece49f235197216e031bb156901c11e744
Revert "Remove novaclient.v1_1 module"

This reverts commit 0cd58123be5e2c2bc2f1233a3e15fb35b2ead1a0

This is a backward incompatible change that is blocking our
ability to release 2.34 for 63c7a57, which was a revert for
another backward incompatible change (2961e82) which slipped
into 2.33.

Once we release 2.34 with the bug fixes, we'll revert this
revert and release that as novaclient 3.0.

Change-Id: If79d37df3e99004897f1bb2d8564d2f1d8aa0978