Updated openstack/openstack Project: openstack/openstack-manuals  3947b5496e0ef231b8e1993d9baf9e6b3dbb8acb
[install] Workaround for RDO lacking PyMySQL

Upstream uses the PyMySQL library in Liberty, but RDO
does not provide a package for it. Installations using
RDO packages must revert to the python-MySQL library
for any service that uses an SQL database which requires
changing the database connection string from
'mysql+pymysql://' to 'mysql://'.

As a workaround, this patch adds conditionals for each
SQL database connection string. After merging, we can
publish the RDO version of the installation guide.

Change-Id: If2423fcfbc08d028d35731aee2896a5bd45c2fb4
Implements: bp installguide-liberty
Closes-Bug: #1501991