Updated openstack/openstack Project: openstack/nova  314225379b7eb4cf713938b1c691fead2126c7f0
Remove SoftDeleteMixin from NovaBase

The mixin is re-added to every class that currently inherits from
NovaBase, so there is no impact to existing classes.

The advantage is that new tables which inherit from NovaBase will not
get the mixin and won't need to have the columns for soft deletion. It
also opens the door to gradually removing the SoftDeleteMixin on a
case-by-case basis.

bp no-more-soft-delete

Change-Id: Id6b37f9a375b383f689f0e23abca7106c9872596