Updated openstack/openstack Project: openstack-infra/project-config  0db8d39fcc5b9f9e2d76d18dcd89d3fbf3a883b9
grenade-multinode: parity with ncpu jobs

We now have fixes that demonstrate the multinode grenade tests working
reasonably well. Our end game is replacing the partial-ncpu job. We
should run this non-voting job everywhere that the partial-ncpu is to
build up enough data to feel comfortable with the voting and flipping
to this job.

Depends-On: I5c899779866ebdf492d474a625eb396273bb0259

Change-Id: I4eefa772b5c9e95608a1ff4e49fc5af60bd0b1a4