Updated openstack/openstack Project: openstack/horizon  e06cf5f38ccba8499bc766a0342cd3713031effb
Changes cursor to hand/grab in network topology.

The network topology view can get rather large and off screen.
When it gets off screen, there is no scroll bars but instead you
can grab the topology and move it around like a map. This ability
is not very clear/intuitive with the default/arrow cursor. This
patch changes the cursor in the topology view to a hand grab cursor
to make this ability more obvious and intuitive.

Closes-Bug: #1503028

Change-Id: I4cc50b4881642e57605fea86a522cd88f1cb7d48