Updated openstack/openstack Project: openstack-infra/project-config  26b484d50c8796b780df3d9ff099a84fb2ffb50c
Move nodepool.yaml from system-config to project-config

Nodepool.yaml.erb in system-config no longer contains any secrets.
Let's move it to the project-config repo.

Add the nodepool validation jobs and tox environment, initially
as non-voting until it is boot strapped.

The version of nodepool.yaml proposed matches the version in system-config
depends-on patch. A follow-up patch will be submitted with any deltas needed
when the needed-by patch is ready to merge. Until that happens, this nodepool.yaml
file won't be used and is safe to merge.

Make both gate-project-config-nodepool and gate-system-config-nodepool
non-voting to aid in bootstrapping this change.

The nodepool logging template file and associated tool that generates the file
will remain in system-config. In the short term, updates to nodepool.yaml in
this repo may require a related change in system-config to update the logging
configuration. In the longer term, nodepool will be updated to automatically
log image creations without needing a customized logging configuration.

Needed-By: I2b45a7145805368b1598d3a3e8a94f0e4eb8cf2d
Depends-On: Ib045c104daac15360e62f46e3e5683c7a387bc96
Depends-On: I89207d100eb4b6bbb502a6ed38831f49e4b29096
Change-Id: I473a1b78acdb035eb379394a7ed5f771434dc942