Updated openstack/openstack Project: openstack/nova  e1666a68ee491c6d6b37dcc9fb5fc3adffe4c446
Revert "Implement online schema migrations"

This reverts commit 2f743b2f20a559c5d2e05aa716b9137e928eb017.

At the Mitaka design summit we discussed removing this as it is
not complete, not under development, and not even an active
experiment for anyone. Since this being in the tree is confusing
people and known to be dangerous, this removes it from the tree.

Conflicts:
	nova/cmd/manage.py
	nova/db/sqlalchemy/migration.py
	nova/exception.py
	nova/tests/unit/db/test_migrations.py

Change-Id: Icae28ceee3ec975c907d73b95babab58dcb30c23