Updated openstack/openstack Project: openstack-infra/system-config  90aeaa21c4ae7bc9d7d311d7e55319a3a2205e73
Add periodic tempest jobs to subunit2sql collection

This commit expands the test result collection to also include
periodic jobs. The data rate for this is relatively low and there
shouldn't be any issues with data purity because the jobs only
run with merged code, so the tests should pass.

Change-Id: I29f930c15fff124cb2c4934a84a999d3ce1a18cb