Updated openstack/openstack Project: openstack/glance-specs  f21c44511d466ea8d306a7437e4ecf2cb6b0bb7d
Add Glance priorities for Mitaka

For this cycle, I believe we should be more strict on what is/isn't a
priority. Considering the latest feedback from the community and our
current team, we ought to be very careful.

This list does not exclude other works. It just clarifies what
the teams priorities are and where most of the time should probably go.

Also, this list does not define features and it doesn't replace specs.
Specs *must* be written and they'll be referenced in this list.

Change-Id: I312d0dcd815da10f083f2fb2a673f7c192c231f4