Updated openstack/openstack Project: openstack-infra/project-config  d6ef7eebabd7185f16577f40b451602ce462d504
Stop running noop against fuel-web

Fuel web has real jobs on some branches and noop on others. The problem
with this is it slows down everyone using noop in the gate because they
share a queue with fuel-web which runs jobs.

Remove the noop job from fuel-web entirely and force them to run pep8 on
the stable branches. This is largely arbitrary but it is trivial to get
a passing pep8 job on old code (just ignore the pep8 rules that fail) so
should be a good choice for them.

Change-Id: I9c77e84c089d5623dcdb047b2402028db77096df