Updated openstack/openstack Project: openstack/ironic-specs  fbbb9434aa37a8720cfbb85c74ceedffea2f00eb
Manual cleaning

Ironic already provides support for automated cleaning. This
specification describes support for manual cleaning, including an
API for operators to specify a list of clean steps to perform on a
node from the MANAGEABLE state.

Clean steps that are destructive and long running such as
configuring RAID or doing burn in, are good candidates for
manual cleaning instead of automated cleaning.

This feature was formerly called 'zapping'.

Change-Id: Iea975cfc2effc2d8be186294b88d85d8f2ace7b2
blueprint: manual-cleaning