Updated openstack/openstack Project: openstack/tempest  0d4493512b2aedb88e4b7aa03a868b72e795f593
Switch to using pbr version for cli version string

This commit switches from using the hard coded version strings in the
cliff main app to using the version string generated pbr. This means
that the version returned from --version on the tempest cli will
actually reflect the version of the code you're using.

Change-Id: I74b2a2c4f1cf9b4a6639c35545b3256ea94cd961