Updated openstack/openstack Project: openstack/keystone  50a2fa5c17879feb76b38a3509dbb08a6860cc67
Promote an arbitrary string to be a docstring

In https://review.openstack.org/#/c/229857/ , Lance pointed out that the
string here was not being caught by pep257, and that turned out to be
because it wasn't actually a docstring at all.

This appears to have been intended as a module-level docstring, but
because it's not the first element in the module, it's actually just a
regular string, and thus not validated against PEP257.

However, rather than promoting this to be a module-level docstring, the
actual content refers to the Saml2 class itself (going so far as to note
the class inheritance and class path for configuration), and thus it
makes sense to make it a class-level docstring.

Change-Id: Ie6250613291c07d4c1a41104453a590fc3a65382