Updated openstack/openstack Project: openstack/tempest  c2e6a88831b9982cb7994aaaab6f0bdf235895e0
Add plugin cookiecutter in documentation

Since we have a official cookiecutter project we should
rely on it in the documentation. All titles are moved a bit
since only the usage of the cookiecutter and the entry point
is needed to create a working test project. All other information
is just additional and not mandatory for plugin creation.

Change-Id: Ia04d2ea747f19d7b1d272bd7664ff05f61c7c88c