Updated openstack/openstack Project: openstack-infra/shade  750701380e65805639a76e1fc459e717496a8662
Add docs for create_server

create_server does a lot of passthrough to nova create_server, which is
great, but means it's tough to know what you need to pass in to shade
when reading the docs. List and document the parameters that shade is
supporting. Continue to support **kwargs pass through for others that
could be used but which shade doesn't know about.

Change-Id: I97cbb50f027e58766015b768a5da9c853b8c29fe