Updated openstack/openstack Project: openstack/oslo.middleware  b4ad4f9073136298294ee9ba10c23ebb2f77a705
Revert "Add 'X-Auth-Token' default header to CORS middleware"

The existence of this header is not tested, and therefore can
create fragility downstream as operators deploying an API
may not realize that this setting is needed. If the upstream
library is ever updated to remove this, all downstream projects
will silently fail.

Furthermore, the addition of openstack-specific headers as
defaults in this library makes oslo_middleware less useful
outside of openstack.

The correct way to address this issue is to enable latent
configuration from paste.ini. This way, each service can
include their default headers in their packaged paste file.

This reverts commit 562173aeb7839eaca6274531b0e91147941af6df.

Change-Id: I401011ed055fea49268686deb5c25f02d2ca6164