Updated openstack/openstack Project: openstack-infra/puppet-gerrit  6566506041bb7d6cd1e8bddc837879e2c5d363e1
Allow setting the number of threads for indexing Gerrit changes

Indexing Gerrit changes can be a time consuming operation.  This
change allows us to set the number of threads to use for indexing
operation which will hopefully make the indexing faster.  If it's
not set Gerrit will (by default) use 1 thread for this operation.

referenced in index.threads section:
 https://review.openstack.org/Documentation/config-gerrit.html

Change-Id: I3aa1909d3edc8aa52d026b3f8e0dfafac4e14f6d