Updated openstack/openstack Project: openstack/cinder  b92cd85cb905898fd51815f178f5203a3df77526
Fix calling delete_zones method with a wrong argument

Add a new zone with cisco zonemanager driver. If the added zone
name is the same as the existed name, the function delete_zones
will be called. But it fails because of passing a argument with
incorrect type. This patch fixes this error.

Co-Authored-By: Al Lau <alau2@cisco.com>
Closes-Bug: #1496283
Change-Id: Ib340b6828cfeb2bb318db23971268c7974ecc9f6