Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  4408bca1edcf33d781cec1da5ca4df78703a14e0
cmd.py: tweak plugin information querying logic

We don't want to unconditionally disable plugin querying just because
--conf is not given. That would mean that the query_plugins_info field
in the default paths would always be ignored, even if set to True.

This patch tweaks the logic so that we remember in options.conf the
actual name of the config file that we read from, whether it be from
--conf or from one of the default paths.

This way, the `test` command will still work out of the box, as commit
f68831b envisioned. But query_plugins_info is now checked if the user
has a configuration file (whether it be a default one, or one specified
through --conf).

Change-Id: Id71ad207f0419f7cc52eca174dd859d4182d7250
Story: 2000387