Updated openstack/openstack Project: openstack-infra/system-config  3493be19160b835f6af08af07bae8e61764490fd
Add some in-tree ansible group vars

As we're using these roles, we'll want to pass potentially different
values to different of our hosts over time. For instance, we may want to
set the jenkins servers to start using puppet apply before we get all
the hosts there. Since we run most of the hosts in a big matching
mechanism, the way we can pass different input values to each host.

Change-Id: I5698355df0c13cd11fe5987787e65ee85a384256