Updated openstack/openstack Project: openstack-infra/shade  93d8b79900435e2f3b1755963e4a465cb6d4cdbf
Don't warn on configured insecure certs

If we have to connect to bogus certs and we know this and we've
configured 'verfiy=False' meaning that we've explicitly indicated
that we want to turn off cert verification, then spamming our logs
with warnings that the certs are invalid is a bit lame. We know. We're
not happy about it - but at least one of our providers has bad certs. We
wish we had the power to change it, but we don't. We already feel bad.
STOP YELLING AT US.

Change-Id: I08d816bcc1685fc9ed3bfcf2e3e8300859059903