Updated openstack/openstack Project: openstack/ceilometer  ccac2a3274522c4901012a615da8e3b5769fb795
Wrong result is returned when call events getting API

For current implementation, events getting API only support 'eq' query
operation when query field is the one of 'event_type', 'message_id',
'start_timestamp' and 'end_timestamp'. But there is a problem, if the
query operation wasn't specified as 'eq'(e.g. 'ne'), the returned result
is still the same as the query operation was specified as 'eq'.

This patch add a check for this situation: if the query field is the one
of 'event_type', 'message_id', 'start_timestamp', 'end_timestamp', and
the query operation user specified is not 'eq', then a client side error
will be thrown.

Also, corresponding unit test case is added.

Change-Id: I4e4b127045de6e933281d9289271af891c3c80fe
Closes-Bug: #1511592