Updated openstack/openstack Project: openstack-infra/project-config  be7dfc521f82f4e4fcf2e3a4dcac360b362324e9
Rename projects akanda-* -> astara-*

The project-formely-known as Akanda has moved to openstack/ and is to be
renamed Astara. This is the final phase of the project move+rename.

This patch renames:

    - openstack/akanda-rug -> openstack/astara
    - openstack/akanda-appliance -> openstack/astara-appliance
    - openstack/akanda-horizon -> openstack/astara-horizon
    - openstatck/akanda-neutron -> openstack/astara-neutron

It also renames gerrit groups:
    - akanda-core -> astara-core
    - akanda-release -> astara-release

As part of the rename, it also retires two unused repos to openstack-attic:

    - akanda
    - akanda-appliance-builder

Jenkins jobs and zuul layouts updated accordingly.

Change-Id: I3d64946d00cf7fd9c8dce725c56b83e256a6fd11