Updated openstack/openstack Project: openstack/swift  7242c4a006c39f6f5a3e53607dc5e2998328b3c2
Fix missing *-replicator conf sections in deployment guide

The doc for these sections was missing because of an rst error - the
source is there in rst file but didn't make it into the html output.

Add doc for per_diff and max_diffs in account and container doc sections.

Also, fix a bunch of other sphinx build errors and most of the warnings.

Change-Id: If9ed2619b2f92c6c65a94f41d8819db8726d3893