Updated openstack/openstack Project: openstack/nova  58f5eb5875ba5624f725560128300596eef75e15
Add debug logging for when boot sequence is invalid in _validate_bdm

The cells job is failing to boot from volume with a bdm v2 request like:

"block_device_mapping_v2": [{
    "destination_type": "volume",
    "boot_index": 0,
    "uuid": "553083ac-40f2-4225-ac4d-a1d021eb7fb1",
    "source_type": "volume",
    "delete_on_termination": true
}]

Given it seems there should only be one BDM in the list and it's boot
index is 0, we shouldn't fail on the boot sequence validation, so add
some debug logging when we hit that failure.

Also copy some wording out of the 'Block Device Mapping in Nova' section
of the devref to add as code comments for the validation happening so we
have context.

TODO(mriedem): track down ndipanov to figure out why _subsequent_list
omits the last element of the list that it's processing, we need to doc
that logic in the code for maintainability.

Change-Id: I8adc94f3c93c149689fbef424665fb44ac573819
Related-Bug: #1489581