Updated openstack/openstack Project: openstack-infra/shade  d65cfbb7908f5b12d1acdce972aa3a44bf86bb6c
Skip an extra unneeded server get

When we're attaching a floating ip, the wait loop gets the most up to
date server dict as part of waiting. In that case, there is no need to
do an additional get. If we're NOT waiting, it's also pointless to do
a single get, as it's not guaranteed to have the updated info.

Change-Id: I84bc54ce32a7d5bf10078027fdb29fab5e84bd39