Updated openstack/openstack Project: openstack-infra/project-config  6f54d56ce8108ce5f7f799e985afaea858c9beec
Make Identity API v3 only job default

There is an experimental gate job for Identity API
v3 only for DevStack and Tempest repositories,
where the Identity v2 API is disabled.

That job aims to verify that Tempest tests can run
with the described configuration, and consequently
services are compatible with the Identity API v3.

In addition, there are two job templates for
running Tempest tests (named *keystonev3-full)
using the Identity API v3 that have the same goal.

Now that the Identity API v3 only job is passing
(see change #213430 as an example), the other two
jobs can be removed in favor of it.

Change-Id: Ie596494df384de024ddeaadbd641f4b93dce30e0