Updated openstack/openstack Project: openstack/tripleo-heat-templates  f453291db56b1504c4e0c0e5df96061ed0950775
Cisco Nexus ML2 MD: Need unique key for server level keys in hieradata

Due to a limitation in the puppet version used in RHEL7 there is no simple
way to scope a 2nd level hiera hash key with the create_resources + defined
types pattern.  Lack of the .each method support prior to puppet 4.0 is the
problem here.  This template change works around the problem by explicitly
adding the hostname to the hieradata for a server under a nexus switch.

The duplicate server names under different switches is needed for vPC
config scenarios.

Closes-bug: #1506546
Change-Id: I03b866fb440e968c9f86ae93942b687e7165a065