Updated openstack/openstack Project: openstack-infra/irc-meetings  a007531ce242d15197b0d886621227712abac338
Replace hyphens with underscores for meeting_id

Update the YAML files to use underscores instead of hyphens for the
'meeting_id' field.

Also handle 'meeting_id' fields with hyphens by converting them to
underscores when generating the URL to the Logs. This will handle
possible future commits where a 'meeting_id' is added with a hyphen.

Change-Id: I7f6574c282e3bcf06c7916334a10332ef8d0cbae