Updated openstack/openstack Project: openstack/neutron  566e92f42d30c381d69934740535c2e6a6a61dcc
Delete fipnamespace when external net removed on DVR

The fipnamespace is associated with an external network
on a given node. In the case of DVR there is just one
single FIP namespace for a given node.

We have seen some race conditions in the agent for creation
and deletion of the fip namespace. See the bug report for
details on the failure.

So in order to address this race condition and make the
code more stable, we will be cleaning up the fip namespace
only when an external network is removed.

The server will be sending a rpc notification message to
the agent to cleanup the fip namespace when the external
net is removed.

This patch address the above mentioned issue by not constantly
deleting and creating the fip namespace.

Closes-Bug: #1501873
Change-Id: I86869f66d4afffad7db09942578b1a456a9bd418