Updated openstack/openstack Project: openstack/nova  e52d236a3f1740997890cad9d4726df01d5a7e5d
cells: add debug logging to bdm_update_or_create_at_top

The method is racy and we need to know which conditions we're hitting
when we fail, which is happening sometimes with the
test_create_ebs_image_and_check_boot test in Tempest.

We fail because the image snapshot metadata has duplicate BDM entries in
it so that when we boot an instance from that image, there are multiple
BDMs with the same boot_index and nova.compute.api.API._validate_bdm
throws up.

Related-Bug: #1489581

Change-Id: I8058e0c16e0fd9b54d4a46e86e0fdede3da55777