Updated openstack/openstack Project: openstack/ceilometer  2bddfa28ad8ee9c04873ffea606287b9a6cd04c7
Factorize yaml loading of declarative stuffs

This change continues to remove code duplication by factorizing the
yaml file loading.

This also homogenizes the error handling.
The precision of error message will be the same for all.
When the yaml file is incorrectly formatted, we now raise an exception,
instead of sometimes raising and sometimes loading an empty definitions,
that can generated an unwanted behavior of collector and notification agents.

Change-Id: I9ae8f519472f1ae2a14e61028a427ba6f3b0d1f3