Updated openstack/openstack Project: openstack-infra/shade  2ffbd8a94612d37c052e4ddc2ee8e2cc6dc14f77
Remove another extraneous get for create_server

The pattern of doing a get after a create is one we've been doing in
Infra for quite a while. However, if we're going to spin off a wait
loop, there is no need for this get, as the only thing we need in the
loop is the server id - which we already have. OTOH, if we're not going
to wait, this is still a useful get to do. Lucky for us, python has a
feature know as "if" which makes expressing this easy.

Change-Id: Ida354fb8fc772325a2fcf8e4ef1dcf3841176e2b