Updated openstack/openstack Project: openstack/glance  52623b9edd23bddedd370cc0884eebc660e4fb2b
Allow owner to be set on image create

This change means that on an image create a admin can set the owner
for the image. I believe this was accidently missed as the owner can be
changed in other ways (and there is a command for it in glance client).

Change-Id: I8d572a070bbb04dccdd051b8e0ad199c5754746e
Closes-bug: 1499368