Updated openstack/openstack Project: openstack/nova  6f199c413d2c2775dd634a7497185bce04b92236
Remove reservation_id from the logs when a schedule fails

As discussed, reservation_id is a compatible layer for EC2 and thus is persisted,
but there are no filters which are using it. Since there are very little benefits
of giving it to the operator (because the operator can find a request by other ways)
and since it would mean we should persist that too in the RequestSpec object,
we prefer to remove that from the log information exposed to the operator.

Change-Id: Ie8f050a1b3653fe1a6d901b3e4020158d63b9b8a
Partially-Implements: blueprint request-spec-object-mitaka