Updated openstack/openstack Project: openstack/neutron-vpnaas  aa5990954a4624f3f00fd81c0262284167cd9dfb
VPNaaS: Remove unneeded test

Endpoint groups for VPN, use subnet CIDRs for local endpoints. This test
was added to address the case where a subnet is modified, while being
used by an endpoint group.

However, further investigation indicates that the the subnet CIDR field
is read-only, so it cannot be updated, when in use by an endpoint group.

As a result, this test case is not needed.

Change-Id: I87262949d0e49d42346fae04ab7f6fb989845c47