Updated openstack/openstack Project: openstack/neutron  78744ae36784972d03eee5140807a63cd9130644
Ensure to decode bytes or fail

The commit 048316e98114145ad19285cdeedd0040b59a66f8 introduces the
pattern:

  if isinstance(line, bytes):
    try:
      line = line.decode(encoding='utf-8')
    except UnicodeError:
      pass
  # concat line with a string

which is not working in PY3K if an UnicodeError is raised because line
is (silently) not decoded and concatened to a string.

This change ensures to return a text object or to raise an error.

Closes-Bug: #1503415
Blueprint: neutron-python3
Change-Id: I16b8013f33aa3efad65be8040d3210120e047bbd